requirements.txt
numpy
torch
streamlit
pillow
soundfile
pyyaml
